requirements.txt
Flask==3.1.0
psycopg2-binary==2.9.7
pytest==7.4.2
python-dotenv==0.19.0
